test tutorial 2 and add readme

+++ mvl_challenge/utils/image_utils.py
@@ -97,11 +97,11 @@ def draw_uncertainty_map(sigma_boundary, peak_boundary, shape=(512, 1024)):
     for u,v, sigma in  zip(peak_boundary[0, :], peak_boundary[1, :], sigma_boundary):
         sigma_bon = (sigma / np.pi) * shape[0]
 
-        sampled_points = np.random.normal(v, sigma_bon, 512).astype(np.int)
+        sampled_points = np.random.normal(v, sigma_bon, 512).astype(np.int16)
         sampled_points[sampled_points >= H] = H-1
         sampled_points[sampled_points <= 0] = 0
         
-        u_point = (np.ones_like(sampled_points) * u).astype(np.int)
+        u_point = (np.ones_like(sampled_points) * u).astype(np.int16)
         img_map[sampled_points, u_point] = 1
        
     img_map = skimage.filters.gaussian(
